Separate publish.gradle.kts from build.gradle.kts
